Fixed test organization
