+++ core/shared/src/test/scala/fs2/hashing/HashingSuite.scala
@@ -20,17 +20,16 @@
  */
 
 package fs2
+package hashing
 
-import cats.effect.IO
+import cats.effect.{IO, Resource}
 import cats.syntax.all._
 import org.scalacheck.Gen
 import org.scalacheck.effect.PropF.forAllF
 
-import hash._
+class HashingSuite extends Fs2Suite with HashSuitePlatform with TestPlatform {
 
-class HashingSuite extends Fs2Suite with HashingSuitePlatform with TestPlatform {
-
-  def checkDigest[A](h: Pipe[IO, Byte, Byte], algo: String, str: String) = {
+  def checkDigest[A](h: Resource[IO, Hash[IO]], algo: String, str: String) = {
     val n =
       if (str.length > 0) Gen.choose(1, str.length).sample.getOrElse(1) else 1
     val s =
@@ -42,27 +41,32 @@ class HashingSuite extends Fs2Suite with HashingSuitePlatform with TestPlatform 
             acc ++ Stream.chunk(Chunk.array(c))
           )
 
-    s.through(h).compile.toList.assertEquals(digest(algo, str))
+    s.through(Hashing[IO].hashWith(h)).compile.toList.assertEquals(digest(algo, str))
   }
 
   group("digests") {
-    if (isJVM) test("md2")(forAllF((s: String) => checkDigest(md2, "MD2", s)))
-    test("md5")(forAllF((s: String) => checkDigest(md5, "MD5", s)))
-    test("sha1")(forAllF((s: String) => checkDigest(sha1, "SHA-1", s)))
-    test("sha256")(forAllF((s: String) => checkDigest(sha256, "SHA-256", s)))
-    test("sha384")(forAllF((s: String) => checkDigest(sha384, "SHA-384", s)))
-    test("sha512")(forAllF((s: String) => checkDigest(sha512, "SHA-512", s)))
+    if (isJVM) test("md2")(forAllF((s: String) => checkDigest(Hashing[IO].create("MD2"), "MD2", s)))
+    test("md5")(forAllF((s: String) => checkDigest(Hashing[IO].md5, "MD5", s)))
+    test("sha1")(forAllF((s: String) => checkDigest(Hashing[IO].sha1, "SHA-1", s)))
+    test("sha256")(forAllF((s: String) => checkDigest(Hashing[IO].sha256, "SHA-256", s)))
+    test("sha384")(forAllF((s: String) => checkDigest(Hashing[IO].sha384, "SHA-384", s)))
+    test("sha512")(forAllF((s: String) => checkDigest(Hashing[IO].sha512, "SHA-512", s)))
   }
 
   test("empty input") {
-    Stream.empty.covary[IO].through(sha1).compile.count.assertEquals(20L)
+    Stream.empty
+      .covary[IO]
+      .through(Hashing[IO].hashWith(Hashing[IO].sha1))
+      .compile
+      .count
+      .assertEquals(20L)
   }
 
   test("zero or one output") {
     forAllF { (lb: List[Array[Byte]]) =>
       val size = lb
         .foldLeft(Stream.empty.covaryOutput[Byte])((acc, b) => acc ++ Stream.chunk(Chunk.array(b)))
-        .through(sha1[IO])
+        .through(Hashing[IO].hashWith(Hashing[IO].sha1))
         .compile
         .count
       size.assertEquals(20L)
@@ -74,7 +78,7 @@ class HashingSuite extends Fs2Suite with HashingSuitePlatform with TestPlatform 
       .range(1, 100)
       .covary[IO]
       .flatMap(i => Stream.chunk(Chunk.array(i.toString.getBytes)))
-      .through(sha512)
+      .through(Hashing[IO].hashWith(Hashing[IO].sha512))
     for {
       once <- s.compile.toVector
       oneHundred <- Vector.fill(100)(s.compile.toVector).parSequence
